+++ test/support/workflow_live_helpers.ex
@@ -253,7 +253,6 @@ defmodule Lightning.WorkflowLive.Helpers do
       %{jobs: []},
       %{jobs: [%{id: id, name: name}]}
     )
-    |> Jsonpatch.Mapper.to_map()
     |> List.first()
     |> Lightning.Helpers.json_safe()
   end
@@ -316,7 +315,6 @@ defmodule Lightning.WorkflowLive.Helpers do
         "project_id" => project.id
       }
     )
-    |> Jsonpatch.Mapper.to_map()
     |> Enum.map(&Lightning.Helpers.json_safe/1)
   end
 
@@ -327,7 +325,7 @@ defmodule Lightning.WorkflowLive.Helpers do
 
     view
     |> element(
-      ~s{input[name='workflow[jobs][#{idx}][#{field}]'] ~ .error-space [data-tag="error_message"]},
+      ~s{div[phx-feedback-for="workflow[jobs][#{idx}][#{field}]"] .error-space [data-tag="error_message"]},
       error
     )
     |> has_element?()